requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
pyarrow==15.0.2
fastparquet==2024.5.0

# PJM data access
gridstatus==0.23.0

# MLOps core
xgboost==2.1.0
mlflow==2.15.0
feast==0.42.0
redis==5.0.4
boto3==1.34.160
pydantic==2.7.1
python-dotenv==1.0.1

# API
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Data quality
great_expectations==0.18.15

# Testing & dev
pytest==8.2.0
black==24.8.0
flake8==7.1.0
